refactor(gh-pages): move all nav components into the same source directory
