+++ pages/src/localdev.less
@@ -18,16 +18,11 @@
 @import "./page/sidebar-layout.less";
 @import "./page/landing-layout.less";
 
-@import "./header/logo.less";
-@import "./header/header.less";
-@import "./footer/footer.less";
-@import "./footer/footer-compact.less";
+@import "./navigation/navigation.less";
 @import "./related/related.less";
 @import "./marquee/marquee.less";
 @import "./collection-grid/collection-grid.less";
 @import "./benefits/benefits.less";
-
-@import "./sidebar/sidebar.less";
 
 @import "../../src/modules/all.less";
 @import "../../src/modules/draft/all.less";
