+++ pages/src/navigation/footer/footer-compact.less
