+++ pages/src/navigation/footer/footer.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .footer {
 
@@ -159,7 +159,6 @@
     }
 
     &-accordion {
-      border-top: none;
       text-align: start;
       flex: 1 0 0px;
       border: none;
