+++ pages/src/navigation/header/header.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 .header {
   z-index: 10;
