+++ pages/src/navigation/navigation.less
@@ -1,0 +1,5 @@
+@import "./header/header.less";
+@import "./header/logo.less";
+@import "./sidebar/sidebar.less";
+@import "./footer/footer.less";
+@import "./footer/footer-compact.less";
