+++ pages/src/navigation/navigation.ts
@@ -1,0 +1,1 @@
+export * from './sidebar/sidebar';
