+++ pages/src/navigation/sidebar/sidebar-content.less
@@ -1,3 +1,5 @@
+@import (reference) "../../common/variables.less";
+
 esl-d-sidebar {
   .sidebar-nav {
     width: 100%;
