+++ pages/src/navigation/sidebar/sidebar-seccondary-nav.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 esl-d-sidebar {
   .sidebar-nav-secondary {
