+++ pages/src/navigation/sidebar/sidebar.less
@@ -1,4 +1,4 @@
-@import (reference) "../common/variables.less";
+@import (reference) "../../common/variables.less";
 
 esl-d-sidebar {
   display: block;
