+++ pages/src/navigation/sidebar/sidebar.ts
@@ -1,11 +1,11 @@
-import {prop} from '../../../src/modules/esl-utils/decorators/prop';
-import {bind} from '../../../src/modules/esl-utils/decorators/bind';
-import {ready} from '../../../src/modules/esl-utils/decorators/ready';
-import {attr} from '../../../src/modules/esl-base-element/decorators/attr';
-import {ESLToggleable} from '../../../src/modules/esl-toggleable/core/esl-toggleable';
-import {ESLMediaQuery} from '../../../src/modules/esl-media-query/core/esl-media-query';
+import {prop} from '../../../../src/modules/esl-utils/decorators/prop';
+import {bind} from '../../../../src/modules/esl-utils/decorators/bind';
+import {ready} from '../../../../src/modules/esl-utils/decorators/ready';
+import {attr} from '../../../../src/modules/esl-base-element/decorators/attr';
+import {ESLToggleable} from '../../../../src/modules/esl-toggleable/core/esl-toggleable';
+import {ESLMediaQuery} from '../../../../src/modules/esl-media-query/core/esl-media-query';
 
-import type {ToggleableActionParams} from '../../../src/modules/esl-toggleable/core/esl-toggleable';
+import type {ToggleableActionParams} from '../../../../src/modules/esl-toggleable/core/esl-toggleable';
 
 interface SidebarActionParams extends ToggleableActionParams {
   /** Change state without animation */
